Add live-updates when publishing plugins
Notify the user of any errors when publishing a plugin and provide
instructions on how to fix those errors via websockets and a
live-updating web page.
