+++ assets/js/plugins_new.js
@@ -10,22 +10,24 @@ abot.PluginsNew.controller = function() {
 		var submitBtn = document.getElementById("submit-btn")
 		submitBtn.setAttribute("disabled", true)
 		var r = document.getElementById("repourl").value
+		var secret = Math.random().toString(36).slice(2)
 		abot.request({
 			method: "POST",
 			url: window.location.origin + "/api/plugins.json",
-			data: { Path: r },
+			data: { Path: r, Secret: secret },
 		}).then(function() {
 			document.getElementById("repourl").value = ""
 			submitBtn.removeAttribute("disabled")
-			// TODO improve this with m.prop
-			abot.successFlash("Success! Your plugin will appear here when processed (usually in seconds).")
-			m.route("/profile", null, true)
+			m.route("/plugins/progress/"+secret, null, true)
 		}, function(err) {
 			console.error(err)
 			document.getElementById("alert-error").classList.remove("hidden")
 			document.getElementById("alert-error-content").innerText = err.Msg
 			submitBtn.removeAttribute("disabled")
 		})
+	}
+	ctrl.props = {
+		error: m.prop(m.route.param("e")),
 	}
 }
 abot.PluginsNew.view = function(ctrl) {
@@ -34,6 +36,11 @@ abot.PluginsNew.view = function(ctrl) {
 		m(".main", [
 			m(".content", [
 				m("h1", "Add plugin"),
+				function() {
+					if (!!ctrl.props.error()) {
+						return m(".alert.alert-error", ctrl.props.error())
+					}
+				}(),
 				m("p", "Manually add to or update a plugin in the itsabot.org index, so it's searchable. You can add any plugin available via `go get`."),
 				m("form", { onsubmit: ctrl.submit }, [
 					m(".content", [
