+++ cmd/migrateup.sh
@@ -2,4 +2,4 @@
 
 set -e
 
-ls db/migrations/up/*.sql | xargs -I{} -- psql -U postgres itsabot -f {}
+ls db/migrations/up/*.sql | xargs -I{} -- psql -U postgres itsabot -h localhost -f {}
